Add files via upload
In this model I use the parkinson telemonitoring data provided by UCI. The objetive is to predict total UPDRS (a scoring measure that meassure the severity of Parkinson Dissease). For this, a set of voice data are using. The goal of the initial proyect was identify Parkinson early symptoms, just with voice recordings signals. For more details see https://archive.ics.uci.edu/ml/datasets/Parkinsons+Telemonitoring and the suggested pappers